Рефакторинг попап карточки
